The Witness: The big dumb refactor (#3007)
