+++ worlds/witness/data/WitnessLogic.txt
