+++ worlds/witness/data/WitnessLogicVanilla.txt
