+++ worlds/witness/data/settings/Door_Shuffle/Complex_Door_Panels.txt
