+++ worlds/witness/data/settings/Door_Shuffle/Complex_Doors.txt
