+++ worlds/witness/data/settings/Door_Shuffle/Obelisk_Keys.txt
