+++ worlds/witness/data/settings/Door_Shuffle/Simple_Additional_Panels.txt
