+++ worlds/witness/data/settings/Door_Shuffle/Simple_Doors.txt
