+++ worlds/witness/data/settings/Door_Shuffle/Simple_Panels.txt
