+++ worlds/witness/data/settings/Exclusions/Vaults.txt
