+++ worlds/witness/data/settings/Postgame/Bottom_Floor_Discard_NonDoors.txt
