+++ worlds/witness/data/settings/Postgame/Challenge_Vault_Box.txt
